Extract TrigBufCache.cpp
